feat: Enhance news load resoures using load-group - EXO-63805
Prior to this change, all news extensions were loaded inside all-bottom-container including the non needed once in all pages.
This PR add the use of load-group mechanism to enhance the resources load and load only what is needed in the current page.
