+++ webapp/src/main/webapp/WEB-INF/conf/news/dynamic-container-configuration.xml
@@ -82,7 +82,7 @@
         </value-param>
         <object-param>
           <name>news-extensions-portlet</name>
-          <description></description>
+          <description>News extensions </description>
           <object type="org.exoplatform.commons.addons.PortletModel">
             <field name="contentId">
               <string>news/NewsExtensions</string>
