+++ webapp/src/main/webapp/WEB-INF/conf/news/nodetypes-templates/news/views/view1.gtmpl
@@ -152,6 +152,7 @@
         activityId: '${activityId}',
         hiddenSpace: ${news.hiddenSpace},
         isSpaceMember: ${news.isSpaceMember},
+        favorite: ${news.isFavorite}
       },
     } """
     def notAuthorizedUserLabel = "";
