+++ webapp/src/main/webapp/news-extensions/activity-extensions/main.js
@@ -13,8 +13,8 @@
   You should have received a copy of the GNU Affero General Public License
   along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
-import * as analyticsExtension from './analytics-extensions/main.js';
+import {initExtensions} from './extensions.js';
 
 export function init() {
-  analyticsExtension.init();
+  initExtensions();
 }
