+++ webapp/src/main/webapp/news-extensions/composer-action-extensions/components/activity/ActivitySwitchToNews.vue
