+++ webapp/src/main/webapp/news-extensions/favorite-drawer-extensions/components/favorites/NewsFavoriteItem.vue
