+++ webapp/src/main/webapp/news-extensions/favorite-drawer-extensions/main.js
@@ -13,8 +13,9 @@
   You should have received a copy of the GNU Affero General Public License
   along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
-import * as analyticsExtension from './analytics-extensions/main.js';
+import './initComponents.js';
+import {initExtensions} from './extensions.js';
 
 export function init() {
-  analyticsExtension.init();
+  initExtensions();
 }
